14_Brand_agent/14. Brand Agent/requirements.txt
openai
python-dotenv
pyyaml
gradio
requests
python-json-logger
